backend/requirements.txt
azure-cognitiveservices-speech
uuid
Flask-Cors
Flask
pybase64